chore(config|constants): rename and move files to better adhere to best practices
- configuration to config, CacheNames
- move ErrorCode to separate package
- move EmployeeMapper to mapper within http
- CorrelationFilter and CorrelationInterceptor to log
