+++ api/src/main/java/com/reliaquest/api/cache/CacheNames.java
@@ -1,7 +1,7 @@
 package com.reliaquest.api.cache;
 
-public final class CacheConstants {
-    private CacheConstants() {}
+public final class CacheNames {
+    private CacheNames() {}
 
     public static final String EMPLOYEES_ALL = "employees:all";
 }
