+++ api/src/main/java/com/reliaquest/api/controller/EmployeeControllerAdvice.java
@@ -1,13 +1,13 @@
 package com.reliaquest.api.controller;
 
-import static com.reliaquest.api.http.ErrorCode.*;
+import static com.reliaquest.api.errors.ErrorCode.*;
 import static com.reliaquest.api.log.LogConstants.MDCKeys.CORRELATION_ID;
 import static com.reliaquest.api.log.LogConstants.PropertyKeys.*;
 
 import com.reliaquest.api.exception.DownstreamUnavailableException;
 import com.reliaquest.api.exception.EmployeeNotFoundException;
 import com.reliaquest.api.exception.EmployeeWithNameAlreadyExistsException;
-import com.reliaquest.api.http.ErrorCode;
+import com.reliaquest.api.errors.ErrorCode;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.ConstraintViolationException;
 import java.net.URI;
