+++ api/src/test/java/com/reliaquest/api/gateway/EmployeeClientTest.java
@@ -1,10 +1,9 @@
 package com.reliaquest.api.gateway;
 
-import com.reliaquest.api.config.PropsConfig;
 import com.reliaquest.api.exception.BadGatewayException;
 import com.reliaquest.api.exception.DownstreamUnavailableException;
 import com.reliaquest.api.exception.EmployeeNotFoundException;
-import com.reliaquest.api.http.CorrelationInterceptor;
+import com.reliaquest.api.log.CorrelationInterceptor;
 import com.reliaquest.api.http.DownstreamErrorHandler;
 import com.reliaquest.api.http.RestClientConfig;
 import com.reliaquest.api.model.CreateEmployeeInput;
@@ -37,7 +36,6 @@ import org.springframework.validation.beanvalidation.MethodValidationPostProcess
     RestClientConfig.class,
     DownstreamErrorHandler.class,
     CorrelationInterceptor.class,
-    PropsConfig.class,
     EmployeeClientTest.MethodValidationConfiguration.class
 })
 class EmployeeClientTest {
