+++ api/src/test/java/com/reliaquest/api/http/EmployeeMapperTest.java
@@ -1,6 +1,7 @@
 package com.reliaquest.api.http;
 
 import com.reliaquest.api.exception.BadGatewayException;
+import com.reliaquest.api.http.mapper.EmployeeMapper;
 import com.reliaquest.api.model.Employee;
 import java.util.UUID;
 import org.junit.jupiter.api.Assertions;
